🐛 8BH6HY supervisor: recover ready stale external episodes
