+++ .agentplane/tasks/202608080551-8BH6HY/pr/github-body.md
@@ -37,8 +37,9 @@ packages/agentplane/src/commands/task/external-agent-purpose.test.ts
  .../external-agent-implementation-authority.ts     |   3 +-
  .../commands/task/external-agent-purpose.test.ts   |  19 ++
  .../src/commands/task/external-agent-purpose.ts    |  11 +
- .../src/commands/task/external-agent-supervisor.ts |  30 +--
- 6 files changed, 400 insertions(+), 17 deletions(-)
+ .../task/external-agent-supervisor-episode.ts      | 116 ++++++++
+ .../src/commands/task/external-agent-supervisor.ts | 113 +-------
+ 7 files changed, 509 insertions(+), 107 deletions(-)
 ```
 
 </details>
