+++ .agentplane/tasks/202608080551-8BH6HY/pr/review.md
@@ -34,8 +34,9 @@ Created: 2026-08-08T05:52:09.195Z
  .../external-agent-implementation-authority.ts     |   3 +-
  .../commands/task/external-agent-purpose.test.ts   |  19 ++
  .../src/commands/task/external-agent-purpose.ts    |  11 +
- .../src/commands/task/external-agent-supervisor.ts |  30 +--
- 6 files changed, 400 insertions(+), 17 deletions(-)
+ .../task/external-agent-supervisor-episode.ts      | 116 ++++++++
+ .../src/commands/task/external-agent-supervisor.ts | 113 +-------
+ 7 files changed, 509 insertions(+), 107 deletions(-)
 ```
 
 </details>
