+++ packages/agentplane/src/commands/task/external-agent-supervisor.ts
@@ -5,8 +5,6 @@ import path from "node:path";
 import {
   advanceSupervisorExecutionEpisodeState,
   completeSupervisorExecutionEpisode,
-  reopenCompletedSupervisorExecutionEpisodeAfterStaleState,
-  startSupervisorExecutionEpisode,
   validateAgentWorkOrderV2,
   type AgentWorkOrderV2,
 } from "@agentplaneorg/core/schemas";
@@ -17,7 +15,6 @@ import { createEvaluatorArtifactPreparationPort } from "../evaluator/evaluator-a
 import type { TaskRouteDecision } from "../shared/route-decision-types.js";
 import {
   createSupervisorEpisodeStore,
-  openSupervisorExecutionEpisode,
   resolveSupervisorExecutionEpisodePath,
   tryAcquireSupervisorExecutionLease,
 } from "../shared/supervisor-execution-episode.js";
@@ -53,6 +50,7 @@ import {
 } from "./external-agent-evaluator.js";
 import { applyExternalImplementationResult } from "./external-agent-implementation-authority.js";
 import { usesExternalImplementationAuthority } from "./external-agent-purpose.js";
+import { recordIssuedExternalAgentEpisode } from "./external-agent-supervisor-episode.js";
 import {
   applyExternalPlanningResult,
   isExternalPlanningResultApplied,
@@ -143,101 +141,6 @@ async function prepareEvaluatorInput(opts: {
   };
 }
 
-async function recordIssuedEpisode(opts: {
-  command: CommandContext;
-  decision: TaskRouteDecision;
-  work_order: AgentWorkOrderV2;
-  work_order_ref: string;
-  purpose: ExternalAgentExchange["purpose"];
-  issue_digest: string;
-}): Promise<void> {
-  const effectRef = `external-agent-issue:${opts.issue_digest}`;
-  for (let attempt = 0; attempt < 3; attempt += 1) {
-    const opened = await openSupervisorExecutionEpisode({
-      git_root: opts.command.resolvedProject.gitRoot,
-      common_git_dir: await resolveCommandGitCommonDir(opts.command),
-      task_id: opts.decision.task.id,
-      task_revision: opts.work_order.task.revision,
-      state_fingerprint_digest: opts.decision.workflowStep.preconditionFingerprint.digest,
-      recover_intent: false,
-    });
-    let journal = opened.journal;
-    if (journal.status === "stopped" && journal.stop?.reason === "stale_state") {
-      const reopened = reopenCompletedSupervisorExecutionEpisodeAfterStaleState({
-        journal,
-        state_fingerprint_digest: opts.decision.workflowStep.preconditionFingerprint.digest,
-      });
-      if (!(await opened.store.compareAndSwap(journal.digest, reopened))) continue;
-      journal = reopened;
-    }
-    if (journal.status === "stopped") {
-      throw new CliError({
-        code: "E_RUNTIME",
-        message: `External-agent supervisor is stopped (${journal.stop?.reason ?? "unknown"}).`,
-      });
-    }
-    if (journal.cursor.phase === "completed") {
-      const advanced = advanceSupervisorExecutionEpisodeState({
-        journal,
-        state_fingerprint_digest: opts.decision.workflowStep.preconditionFingerprint.digest,
-        route_observation: { step_id: opts.decision.workflowStep.id, surface: "task advance" },
-      });
-      if (!(await opened.store.compareAndSwap(journal.digest, advanced))) continue;
-      journal = advanced;
-    }
-    if (journal.cursor.phase === "intent_recorded") {
-      const latest = journal.operations.at(-1);
-      if (
-        latest?.status === "intent" &&
-        latest.precondition_fingerprint_digest ===
-          opts.decision.workflowStep.preconditionFingerprint.digest &&
-        latest.work_order_ref === opts.work_order_ref &&
-        latest.role === opts.work_order.role &&
-        latest.effect_ref === effectRef
-      ) {
-        return;
-      }
-      throw new CliError({
-        code: "E_RUNTIME",
-        message: "Another unresolved external-agent episode already owns this task.",
-      });
-    }
-    if (journal.cursor.phase !== "ready") {
-      throw new CliError({
-        code: "E_RUNTIME",
-        message: "External-agent supervisor is not ready to issue a semantic work order.",
-      });
-    }
-    const started = startSupervisorExecutionEpisode({
-      journal,
-      role: opts.work_order.role,
-      kind: opts.purpose === "quality_review" ? "evaluator_episode" : "agent_episode",
-      operation_identity: {
-        workflow_step_id: opts.decision.workflowStep.id,
-        purpose: opts.purpose,
-        task_id: opts.decision.task.id,
-      },
-      precondition_fingerprint_digest: opts.decision.workflowStep.preconditionFingerprint.digest,
-      authority_ref: `external-agent:${opts.decision.task.id}:${opts.decision.workflowStep.id}`,
-      authority_digest: opts.decision.workflowStep.preconditionFingerprint.digest,
-      work_order_ref: opts.work_order_ref,
-      effect_ref: effectRef,
-    });
-    if (started.status !== "started") {
-      await opened.store.write(started.journal);
-      throw new CliError({
-        code: "E_RUNTIME",
-        message: `External-agent episode could not start (${started.status}).`,
-      });
-    }
-    if (await opened.store.compareAndSwap(journal.digest, started.journal)) return;
-  }
-  throw new CliError({
-    code: "E_RUNTIME",
-    message: "External-agent supervisor changed concurrently while issuing the work order.",
-  });
-}
-
 export async function issueExternalAgentExchange(opts: {
   ctx: CommandCtx;
   command: CommandContext;
@@ -273,7 +176,7 @@ export async function issueExternalAgentExchange(opts: {
       checkout,
     });
     const checkoutCommand = await commandContextForCheckout({ command: opts.command, checkout });
-    await recordIssuedEpisode({
+    await recordIssuedExternalAgentEpisode({
       command: checkoutCommand,
       decision: opts.decision,
       work_order: workOrder,
@@ -328,7 +231,7 @@ export async function issueExternalAgentExchange(opts: {
     updated_at: at,
   };
   await persistExternalAgentExchangeArtifacts({ paths, work_order: workOrder, exchange });
-  await recordIssuedEpisode({
+  await recordIssuedExternalAgentEpisode({
     command: checkoutCommand,
     decision: opts.decision,
     work_order: workOrder,
